chore: bump bb.js timeouts (#6196)
Seem to be hitting timeouts with new bb.js tests. This is a big
worrying, means builds could be interfering with each other, but doing
this for stability

+++ barretenberg/ts/src/barretenberg/common.test.ts
@@ -5,10 +5,12 @@ describe('env', () => {
 
   beforeAll(async () => {
     api = await Barretenberg.new({ threads: 3 });
-  }, 15000);
+  }, 30000);
 
   afterAll(async () => {
-    await api.destroy();
+    if (api) {
+      await api.destroy();
+    }
   });
 
   it('thread test', async () => {
